+++ spec/integration/node/path_spec.rb
@@ -256,6 +256,13 @@ describe "{{path}}" do
 
           expect(errors.first.position).to eq(Curlybars::Position.new(nil, 1, 9, 12))
         end
+
+        it "raises with length 0 when the error has no length" do
+          source = "{{ok}"
+          errors = Curlybars.validate(dependency_tree, source)
+
+          expect(errors.first.position).to eq(Curlybars::Position.new(nil, 1, 4, 0))
+        end
       end
     end
   end