+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* BUGFIX: Error position has 0 as length default value (before was nil)
+
 ### Curlybars 0.9.8 (June 22, 2017)
 
 * Embed length of validation error in position object